+++ src/game/game/game.c
@@ -184,6 +184,15 @@ bool Game_Start(int32_t level_num, GAMEFLOW_LEVEL_TYPE level_type)
         }
         break;
 
+    case GFL_BONUS:
+        Savegame_CarryCurrentInfoToNextLevel(level_num - 1, level_num);
+        Savegame_ApplyLogicToCurrentInfo(level_num);
+        Level_InitialiseFlags();
+        if (!Level_Initialise(level_num)) {
+            return false;
+        }
+        break;
+
     default:
         Level_InitialiseFlags();
         if (!Level_Initialise(level_num)) {
@@ -207,7 +216,11 @@ int32_t Game_Stop(void)
     if (g_CurrentLevel == g_GameFlow.last_level_num) {
         g_Config.profile.new_game_plus_unlock = true;
         Config_Write();
-    } else {
+        g_GameInfo.bonus_level_unlock =
+            Stats_CheckAllSecretsCollected(GFL_NORMAL);
+    }
+
+    if (g_CurrentLevel + 1 < g_GameFlow.level_count) {
         Savegame_CarryCurrentInfoToNextLevel(
             g_CurrentLevel, g_CurrentLevel + 1);
         Savegame_ApplyLogicToCurrentInfo(g_CurrentLevel + 1);
@@ -257,7 +270,7 @@ int32_t Game_Loop(GAMEFLOW_LEVEL_TYPE level_type)
         Stats_GetSecrets();
 
     bool ask_for_save = g_Config.enable_save_crystals
-        && level_type == GFL_NORMAL
+        && (level_type == GFL_NORMAL || level_type == GFL_BONUS)
         && g_CurrentLevel != g_GameFlow.first_level_num
         && g_CurrentLevel != g_GameFlow.gym_level_num;