+++ src/game/savegame.h
@@ -13,7 +13,7 @@
 // creatures, triggers etc., and is what actually sets Lara's health, creatures
 // status, triggers, inventory etc.
 
-#define SAVEGAME_CURRENT_VERSION 3
+#define SAVEGAME_CURRENT_VERSION 4
 
 typedef enum SAVEGAME_VERSION {
     VERSION_LEGACY = -1,
@@ -21,6 +21,7 @@ typedef enum SAVEGAME_VERSION {
     VERSION_1 = 1,
     VERSION_2 = 2,
     VERSION_3 = 3,
+    VERSION_4 = 4,
 } SAVEGAME_VERSION;
 
 typedef enum SAVEGAME_FORMAT {