+++ src/game/stats.h
@@ -2,6 +2,7 @@
 
 #include "global/types.h"
 
+#include <stdbool.h>
 #include <stdint.h>
 
 void Stats_ObserveRoomsLoad(void);
@@ -11,4 +12,5 @@ int32_t Stats_GetPickups(void);
 int32_t Stats_GetKillables(void);
 int32_t Stats_GetSecrets(void);
 void Stats_Show(int32_t level_num);
-void Stats_ShowTotal(const char *filename);
+void Stats_ShowTotal(const char *filename, GAMEFLOW_LEVEL_TYPE level_type);
+bool Stats_CheckAllSecretsCollected(GAMEFLOW_LEVEL_TYPE level_type);