+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - added Linux builds and toolchain
 - added an option to allow Lara to roll while underwater, similar to TR2+ (#993)
 - fixed baddies dropping duplicate guns (only affects mods) (#1000)
+- added the bonus level type for custom levels that unlocks if all main game secrets are found (#645)
 - improved frame scheduling to use less CPU (#985)
 
 ## [2.16](https://github.com/rr-/Tomb1Main/compare/2.15.3...2.16) - 2023-09-20